Don't auto-load default maps
When the plugin was loading, it was setting the default maps for every
buffer whether an audio file had been loaded with TranscibeAudio or not
so that typing, for example, <C-l> in insert mode would insert the text "<plug>(transcribe-seek-forward)" literally.

Worse, because the default mappings map <C-h>, loading the plugin
rendered the backspace key inoperative for any terminal that sends
backspace as ^H.

This changeset  causes the default mappings to be set only the first time :TranscribeAudio is used per buffer.